requirements.txt
# requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
pymongo
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
